fix(core/ui): timer handling for rust layouts
[no changelog]
